update: restructure login page
