+++ frontend/src/helpers/UserAuthApi.js
@@ -18,22 +18,6 @@ const UserAuthApi = {
     return fetch(`${process.env.REACT_APP_API}/register`, options);
   },
 
-  login: (params) => {
-    const options = {
-      method: 'POST',
-      headers: {
-        'Content-Type': 'application/json',
-        'X-Requested-With': 'XMLHttpRequest'
-      },
-      body: JSON.stringify({
-        email: params.email,
-        password: params.password
-      })
-    };
-
-    return fetch(`${process.env.REACT_APP_API}/login`, options);
-  },
-
   logout: (params) => {
     const options = {
       method: 'POST',
