+++ frontend/src/pages/shared/Login/Login.js
@@ -1,16 +1,17 @@
 import Swal from 'sweetalert2';
 import { useState, useEffect } from 'react';
 import { Form, Container, Row, Button } from 'react-bootstrap';
-import { useHistory } from 'react-router-dom';
+import { Link, useHistory } from 'react-router-dom';
 
-import Head from '../components/Head';
-import Footer from '../components/Footer';
-import Navlink from '../components/Navlink';
-import FormTemplate from '../components/FormTemplate';
-import UserAuthApi from '../helpers/UserAuthApi';
-import NavButton from '../components/NavButton';
+import Head from '../../../shared/Head/Head';
+import Footer from '../../../shared/Footer/Footer';
+import Navlink from '../../../shared/Navlink.js/Navlink';
+import FormTemplate from '../../../shared/FormTemplate/FormTemplate';
+import StudentAPI from '../../../api/StudentAPI';
+import NavbarToggle from '../../../shared/NavbarToggle/NavbarToggle';
+import styles from './Login.module.css';
 
-export default function LoginPage() {
+const Login = () => {
   const history = useHistory();
   const [ email, setEmail ] = useState('');
   const [ password, setPassword ] = useState('');
@@ -32,10 +33,10 @@ export default function LoginPage() {
   };
 
   // function to log in a user
-  function loginUser(e) {
+  const loginUser = (e) => {
     e.preventDefault();
 
-    UserAuthApi.login(payload).then(res => res.json()).then(data => {
+    StudentAPI.login(payload).then(res => res.json()).then(data => {
       if (data.token === undefined) {
         Swal.fire({
           icon: 'error',
@@ -61,8 +62,10 @@ export default function LoginPage() {
       <Head title="Login | E-Learning System" />
 
       <nav className="navbar navbar-expand-lg navbar-light bg-white border-bottom border-dark wrapper">
-        <a className="navbar-brand font-weight-bold my-3" href="/"><span className="p-3 rounded-lg border border-dark brandName">E-Learning System</span></a>
-        <NavButton />
+        <Link className="navbar-brand font-weight-bold my-3" to="/">
+          <span className={`p-3 rounded-lg border border-dark ${styles.brandName}`}>E-Learning System</span>
+        </Link>
+        <NavbarToggle />
         <div className="collapse navbar-collapse justify-content-end" id="navbarNavDropdown">
           <ul className="navbar-nav">
             <Navlink />
@@ -73,7 +76,7 @@ export default function LoginPage() {
       <Container className="wrapper">
         <Row className="justify-content-center">
           <Form onSubmit={(e) => loginUser(e)} className="formSize bg-white p-4">
-            <h1 className="my-5 pt-5 text-center font-weight-bold">Login Page</h1>
+            <h1 className={`my-5 pt-5 text-center ${styles.headerFontWeight}`}>Login Page</h1>
             <FormTemplate
               formControlId="formBasicEmail"
               formLabel="Email"
@@ -94,15 +97,17 @@ export default function LoginPage() {
 
             {
               btnActive ?
-                <Button className="bg-primary d-block mx-auto w-100 my-5" type="submit">Submit</Button>
+                <Button className="bg-warning text-dark d-block mx-auto w-100 my-5 border-0 font-weight-bold" type="submit">Submit</Button>
                 :
-                <Button className="bg-primary d-block mx-auto w-100 my-5" type="submit" disabled>Submit</Button>
+                <Button className={`bg-warning text-dark d-block mx-auto w-100 my-5 border-0 font-weight-bold`} type="submit" disabled>Submit</Button>
             }
           </Form>
         </Row>
       </Container>
 
-      <Footer footerClass="text-center border-top pt-5 pb-5 mt-5 wrapper footerCustom"/>
+      <Footer footerClass={`text-center border-top pt-5 pb-5 mt-5 wrapper ${styles.footerClass}`} />
     </>
   );
 };
+
+export default Login;
