+++ frontend/src/shared/Navlink.js/Navlink.module.css
@@ -6,6 +6,6 @@
 .navLink:hover {
   text-decoration: none;
   color: black;
-  font-weight: 700;
+  font-weight: 500;
   border-bottom: 1px black solid;
 }
